Core - Rename CefSharp.Internals.CefRequestWrapper to CefSharp.Request
- Made public
- Added public constructor

Related issue #1721
