+++ CefSharp.Core/CefSharp.Core.vcxproj
@@ -240,7 +240,6 @@
     <ClCompile Include="Internals\CefValueWrapper.cpp" />
     <ClCompile Include="Internals\ClientAdapter.cpp" />
     <ClCompile Include="Internals\CookieVisitor.cpp" />
-    <ClCompile Include="Internals\CefRequestWrapper.cpp" />
     <ClCompile Include="Internals\JavascriptCallbackProxy.cpp" />
     <ClCompile Include="Internals\JavascriptCallbackFactory.cpp" />
     <ClCompile Include="Internals\Serialization\ObjectsSerialization.cpp" />
@@ -251,6 +250,7 @@
     <ClCompile Include="Internals\Serialization\Primitives.cpp" />
     <ClCompile Include="Internals\Serialization\JsObjectsSerialization.cpp" />
     <ClCompile Include="Internals\CefRequestContextHandlerAdapter.cpp" />
+    <ClCompile Include="Request.cpp" />
     <ClCompile Include="RequestContext.cpp" />
     <ClCompile Include="ResourceHandlerWrapper.cpp" />
     <ClCompile Include="Stdafx.cpp">
@@ -307,13 +307,13 @@
     <ClInclude Include="Internals\CefRequestContextHandlerAdapter.h" />
     <ClInclude Include="PostData.h" />
     <ClInclude Include="PostDataElement.h" />
+    <ClInclude Include="Request.h" />
     <ClInclude Include="resource.h" />
     <ClInclude Include="WindowInfo.h" />
     <ClInclude Include="Internals\CefCompletionCallbackAdapter.h" />
     <ClInclude Include="Internals\ClientAdapter.h" />
     <ClInclude Include="Internals\CookieVisitor.h" />
     <ClInclude Include="Internals\CefSharpApp.h" />
-    <ClInclude Include="Internals\CefRequestWrapper.h" />
     <ClInclude Include="Internals\CefJSDialogCallbackWrapper.h" />
     <ClInclude Include="Internals\JavascriptCallbackProxy.h" />
     <ClInclude Include="Internals\JavascriptCallbackFactory.h" />
